添加导入单个基金的蛋卷买入记录

+++ src/main/java/io/renren/modules/etf/danjuan/Data.java
@@ -41,6 +41,6 @@ public class Data {
     private String ttype;
     private long uid;
     private String undo_tip;
-    private long volume;
+    private BigDecimal volume;
 
 }
